extracting cost calculator into separate package with new types package for shared TS types
