+++ examples/basic/package.json
@@ -10,6 +10,7 @@
     "@modelfusion/pinecone": "0.4.0",
     "@modelfusion/serpapi-tool": "0.11.0",
     "@modelfusion/sqlite-vss": "0.4.0",
+    "@modelfusion/types": "workspace:*",
     "@pinecone-database/pinecone": "0.1.6",
     "dotenv": "16.3.1",
     "secure-json-parse": "2.7.0",
