+++ examples/basic/src/experimental/cost/embed-cost-calculation-example.ts
@@ -3,7 +3,6 @@ import { DefaultRun, embed, embedMany, openai } from "modelfusion";
 import {
   OpenAICostCalculator,
   calculateCost,
-  extractSuccessfulModelCalls,
 } from "@modelfusion/cost-calculator";
 dotenv.config();
 
@@ -26,7 +25,7 @@ async function main() {
   });
 
   const cost = await calculateCost({
-    calls: extractSuccessfulModelCalls(run.events),
+    calls: run.getSuccessfulModelCalls(),
     costCalculators: [new OpenAICostCalculator()],
   });
