+++ examples/basic/src/experimental/cost/generate-image-cost-calculation-example.ts
@@ -3,7 +3,6 @@ import { DefaultRun, generateImage, openai } from "modelfusion";
 import {
   OpenAICostCalculator,
   calculateCost,
-  extractSuccessfulModelCalls,
 } from "@modelfusion/cost-calculator";
 
 dotenv.config();
@@ -23,7 +22,7 @@ async function main() {
   });
 
   const cost = await calculateCost({
-    calls: extractSuccessfulModelCalls(run.events),
+    calls: run.getSuccessfulModelCalls(),
     costCalculators: [new OpenAICostCalculator()],
   });
