+++ examples/basic/src/experimental/cost/generate-text-cost-calculation-example.ts
@@ -3,7 +3,6 @@ import { DefaultRun, generateText, openai } from "modelfusion";
 import {
   OpenAICostCalculator,
   calculateCost,
-  extractSuccessfulModelCalls,
 } from "@modelfusion/cost-calculator";
 
 dotenv.config();
@@ -24,7 +23,7 @@ async function main() {
   console.log(text);
 
   const cost = await calculateCost({
-    calls: extractSuccessfulModelCalls(run.events),
+    calls: run.getSuccessfulModelCalls(),
     costCalculators: [new OpenAICostCalculator()],
   });
