+++ packages/@modelfusion-cost-calculator/src/index.ts
@@ -1,5 +1,4 @@
-export * from "./Cost.js";
-export * from "./CostCalculator.js";
-export * from "./SuccessfulModelCall.js";
-export * from "./calculateCost.js";
-export * from "./openai/OpenAICostCalculator.js";
+export * from "./Cost";
+export * from "./CostCalculator";
+export * from "./calculateCost";
+export * from "./openai/OpenAICostCalculator";
