+++ packages/@modelfusion-cost-calculator/src/openai/calculators/transcription.ts
@@ -1,0 +1,24 @@
+import {
+  OpenAITranscriptionModelType,
+  OpenAITranscriptionVerboseJsonResponse,
+  OPENAI_TRANSCRIPTION_MODELS,
+} from "@modelfusion/types";
+
+export const calculateOpenAITranscriptionCostInMillicents = ({
+  model,
+  response,
+}: {
+  model: OpenAITranscriptionModelType;
+  response: OpenAITranscriptionVerboseJsonResponse;
+}): number | null => {
+  if (model !== "whisper-1") {
+    return null;
+  }
+
+  const durationInSeconds = response.duration;
+
+  return (
+    Math.ceil(durationInSeconds) *
+    OPENAI_TRANSCRIPTION_MODELS[model].costInMillicentsPerSecond
+  );
+};
