+++ packages/@modelfusion-types/src/openai/constants/context-window-sizes.ts
@@ -1,0 +1,20 @@
+// https://platform.openai.com/docs/models
+export const OPENAI_CHAT_CONTEXT_WINDOW_SIZES = {
+  "gpt-4": 8192,
+  "gpt-4-0314": 8192,
+  "gpt-4-0613": 8192,
+  "gpt-4-turbo-preview": 128000,
+  "gpt-4-1106-preview": 128000,
+  "gpt-4-0125-preview": 128000,
+  "gpt-4-vision-preview": 128000,
+  "gpt-4-32k": 32768,
+  "gpt-4-32k-0314": 32768,
+  "gpt-4-32k-0613": 32768,
+  "gpt-3.5-turbo": 4096,
+  "gpt-3.5-turbo-0125": 16385,
+  "gpt-3.5-turbo-1106": 16385,
+  "gpt-3.5-turbo-0301": 4096,
+  "gpt-3.5-turbo-0613": 4096,
+  "gpt-3.5-turbo-16k": 16384,
+  "gpt-3.5-turbo-16k-0613": 16384,
+} as const;
