+++ packages/@modelfusion-types/tsconfig.json
@@ -21,6 +21,6 @@
     "skipLibCheck": true,
     "forceConsistentCasingInFileNames": true,
   },
-  "include": ["src/**/*", "../@modelfusion-types/src/SuccessfulModelCall.ts"],
+  "include": ["src/**/*"],
   "exclude": ["node_modules"],
 }
