+++ packages/modelfusion/package.json
@@ -60,8 +60,10 @@
     "lint": "eslint --ext .ts src",
     "clean": "rimraf build dist .turbo node_modules",
     "clean:build": "rimraf build dist",
-    "build": "tsup && pnpm build:copy-files",
+    "build": "tsup --dts && pnpm build:copy-files",
     "build:copy-files": "copyfiles --flat package.json ./README.md ../../LICENSE ./CHANGELOG.md dist",
+    "build:dtsMap": "tsc --emitDeclarationOnly --declaration --declarationMap",
+    "dev": "tsup --watch src --onSuccess \"pnpm build:dtsMap\"",
     "test": "vitest --config vitest.config.js --run src",
     "test:watch": "vitest watch --config vitest.config.js",
     "test:coverage": "vitest run --config vitest.config.js --coverage",
@@ -69,6 +71,7 @@
     "dist": "pnpm clean:build && pnpm lint && pnpm test && pnpm build "
   },
   "dependencies": {
+    "@modelfusion/types": "workspace:*",
     "eventsource-parser": "1.1.1",
     "js-tiktoken": "1.0.7",
     "nanoid": "3.3.6",
