+++ packages/modelfusion/src/index.ts
@@ -1,3 +1,4 @@
+export * from "@modelfusion/types";
 export * from "./core/index";
 export * from "./model-function/index";
 export * from "./model-provider/index";
