+++ packages/modelfusion/src/model-provider/openai/AbstractOpenAIChatModel.ts
@@ -1,4 +1,10 @@
-import { z } from "zod";
+import {
+  openAIChatResponseSchema,
+  openaiChatChunkSchema,
+  OpenAIChatResponse,
+  OpenAIChatChunk,
+} from "@modelfusion/types";
+
 import { FunctionCallOptions } from "../../core/FunctionOptions";
 import { ApiConfiguration } from "../../core/api/ApiConfiguration";
 import { callWithRetryAndThrottle } from "../../core/api/callWithRetryAndThrottle";
@@ -340,106 +346,6 @@ export abstract class AbstractOpenAIChatModel<
     };
   }
 }
-
-const openAIChatResponseSchema = z.object({
-  id: z.string(),
-  choices: z.array(
-    z.object({
-      message: z.object({
-        role: z.literal("assistant"),
-        content: z.string().nullable(),
-        function_call: z
-          .object({
-            name: z.string(),
-            arguments: z.string(),
-          })
-          .optional(),
-        tool_calls: z
-          .array(
-            z.object({
-              id: z.string(),
-              type: z.literal("function"),
-              function: z.object({
-                name: z.string(),
-                arguments: z.string(),
-              }),
-            })
-          )
-          .optional(),
-      }),
-      index: z.number().optional(), // optional for OpenAI compatible models
-      logprobs: z.nullable(z.any()),
-      finish_reason: z
-        .enum([
-          "stop",
-          "length",
-          "tool_calls",
-          "content_filter",
-          "function_call",
-        ])
-        .optional()
-        .nullable(),
-    })
-  ),
-  created: z.number(),
-  model: z.string(),
-  system_fingerprint: z.string().optional().nullable(),
-  object: z.literal("chat.completion"),
-  usage: z.object({
-    prompt_tokens: z.number(),
-    completion_tokens: z.number(),
-    total_tokens: z.number(),
-  }),
-});
-
-export type OpenAIChatResponse = z.infer<typeof openAIChatResponseSchema>;
-
-const openaiChatChunkSchema = z.object({
-  object: z.string(), // generalized for openai compatible providers, z.literal("chat.completion.chunk")
-  id: z.string(),
-  choices: z.array(
-    z.object({
-      delta: z.object({
-        role: z.enum(["assistant", "user"]).optional(),
-        content: z.string().nullable().optional(),
-        function_call: z
-          .object({
-            name: z.string().optional(),
-            arguments: z.string().optional(),
-          })
-          .optional(),
-        tool_calls: z
-          .array(
-            z.object({
-              id: z.string(),
-              type: z.literal("function"),
-              function: z.object({
-                name: z.string(),
-                arguments: z.string(),
-              }),
-            })
-          )
-          .optional(),
-      }),
-      finish_reason: z
-        .enum([
-          "stop",
-          "length",
-          "tool_calls",
-          "content_filter",
-          "function_call",
-        ])
-        .nullable()
-        .optional(),
-      index: z.number(),
-    })
-  ),
-  created: z.number(),
-  model: z.string().optional(), // optional for OpenAI compatible models
-  system_fingerprint: z.string().optional().nullable(),
-});
-
-export type OpenAIChatChunk = z.infer<typeof openaiChatChunkSchema>;
 
 export type OpenAIChatResponseFormatType<T> = {
   stream: boolean;
