+++ packages/modelfusion/src/model-provider/openai/OpenAICompletionModel.ts
@@ -1,3 +1,7 @@
+import {
+  getOpenAICompletionModelInformation,
+  OpenAICompletionModelType,
+} from "@modelfusion/types";
 import { PromptTemplateTextStreamingModel } from "../../model-function/generate-text/PromptTemplateTextStreamingModel";
 import {
   TextStreamingBaseModel,
@@ -13,55 +17,8 @@ import { countTokens } from "../../model-function/tokenize-text/countTokens";
 import {
   AbstractOpenAICompletionModel,
   AbstractOpenAICompletionModelSettings,
-  OpenAICompletionResponse,
 } from "./AbstractOpenAICompletionModel";
 import { TikTokenTokenizer } from "./TikTokenTokenizer";
-
-/**
- * @see https://platform.openai.com/docs/models/
- * @see https://openai.com/pricing
- */
-export const OPENAI_TEXT_GENERATION_MODELS = {
-  "gpt-3.5-turbo-instruct": {
-    contextWindowSize: 4097,
-    promptTokenCostInMillicents: 0.15,
-    completionTokenCostInMillicents: 0.2,
-  },
-};
-
-export function getOpenAICompletionModelInformation(
-  model: OpenAICompletionModelType
-): {
-  contextWindowSize: number;
-  promptTokenCostInMillicents: number;
-  completionTokenCostInMillicents: number;
-} {
-  return OPENAI_TEXT_GENERATION_MODELS[model];
-}
-
-export type OpenAICompletionModelType =
-  keyof typeof OPENAI_TEXT_GENERATION_MODELS;
-
-export const isOpenAICompletionModel = (
-  model: string
-): model is OpenAICompletionModelType => model in OPENAI_TEXT_GENERATION_MODELS;
-
-export const calculateOpenAICompletionCostInMillicents = ({
-  model,
-  response,
-}: {
-  model: OpenAICompletionModelType;
-  response: OpenAICompletionResponse;
-}) => {
-  const modelInformation = getOpenAICompletionModelInformation(model);
-
-  return (
-    response.usage.prompt_tokens *
-      modelInformation.promptTokenCostInMillicents +
-    response.usage.completion_tokens *
-      modelInformation.completionTokenCostInMillicents
-  );
-};
 
 export interface OpenAICompletionModelSettings
   extends AbstractOpenAICompletionModelSettings {
