+++ packages/modelfusion/src/model-provider/openai/OpenAISpeechModel.ts
@@ -1,3 +1,4 @@
+import { OpenAISpeechModelType } from "@modelfusion/types";
 import { FunctionCallOptions } from "../../core/FunctionOptions";
 import { ApiConfiguration } from "../../core/api/ApiConfiguration";
 import { callWithRetryAndThrottle } from "../../core/api/callWithRetryAndThrottle";
@@ -12,36 +13,6 @@ import {
 } from "../../model-function/generate-speech/SpeechGenerationModel";
 import { OpenAIApiConfiguration } from "./OpenAIApiConfiguration";
 import { failedOpenAICallResponseHandler } from "./OpenAIError";
-
-/**
- * @see https://openai.com/pricing
- */
-export const OPENAI_SPEECH_MODELS = {
-  "tts-1": {
-    costInMillicentsPerCharacter: 1.5, // = 1500 / 1000,
-  },
-  "tts-1-hd": {
-    costInMillicentsPerCharacter: 3, // = 3000 / 1000
-  },
-};
-
-export type OpenAISpeechModelType = keyof typeof OPENAI_SPEECH_MODELS;
-
-export const calculateOpenAISpeechCostInMillicents = ({
-  model,
-  input,
-}: {
-  model: OpenAISpeechModelType;
-  input: string;
-}): number | null => {
-  if (!OPENAI_SPEECH_MODELS[model]) {
-    return null;
-  }
-
-  return (
-    input.length * OPENAI_SPEECH_MODELS[model].costInMillicentsPerCharacter
-  );
-};
 
 export type OpenAISpeechVoice =
   | "alloy"
