+++ packages/modelfusion/src/model-provider/openai/TikTokenTokenizer.ts
@@ -1,10 +1,12 @@
 import { Tiktoken } from "js-tiktoken/lite";
 import cl100k_base from "js-tiktoken/ranks/cl100k_base";
+import { OpenAIChatBaseModelType } from "@modelfusion/types";
 import { FullTokenizer } from "../../model-function/tokenize-text/Tokenizer";
 import { never } from "../../util/never";
-import { OpenAIChatBaseModelType } from "./OpenAIChatModel";
-import { OpenAICompletionModelType } from "./OpenAICompletionModel";
-import { OpenAITextEmbeddingModelType } from "./OpenAITextEmbeddingModel";
+import {
+  OpenAICompletionModelType,
+  OpenAITextEmbeddingModelType,
+} from "@modelfusion/types";
 
 export type TikTokenTokenizerSettings = {
   model:
